IndexedDB WPT: Shard slow test
The structured-clone-any.js test is slow as it creates a new database
for each of over a hundred test cases, leading to long (>20s) run
times on debug builds. This would lead to occasional timeouts.

Use the WPT "variant" mechanism and shard the test down to just 20
cases per shard, which drops the test times down to just a few seconds
each (on debug builds).

Bug: 991415
Change-Id: I5fa792ea0042c7ff6ea29d926a459c625958d10d