Add touch-action:none to multi touch tests to prevent firing gestures.
These tests might produce gestures due to the pointer moves that are
being fired, but they are not trying to exercise such behavior. They
currently assume a slop region of 15 pixels as a requirement for
gestures to get fired, but if that region shrinks, these tests will
fail. Adding touch-action:none in the target element's style ensures
that no gesture gets produced.

Bug: 1048893
Change-Id: Iabe6c1dbc9a710a5417aa38290ccf37fd958f198