Make form submissions async in blink
This change implements https://html.spec.whatwg.org/C/#plan-to-navigate
by adding an async step to form submissions which was removed in
https://crrev.com/c/1520353.

This patch also reverts https://crrev.com/c/1850358 which was a
temporary fix for duplicate form submissions caused by sync form
submissions. Now that they are async with this patch again, the
temporary fix is no longer needed.

Bug: 1013385
Change-Id: Ieec850c9684c796ac0582917494ff4dc3ac8f1b1